requirements.lowest.txt
# exactly pinned dependencies to the lowest version regardless of python_version
# see pyptoject file for ranges

packageurl-python == 0.9.0
py-serializable == 0.11.0
importlib-metadata == 3.4.0 # ; python_version < '3.8'
setuptools == 47.0.0
types-setuptools == 57.0.0
toml == 0.10.0
types-toml == 0.10.0
